+++ src/imagedata/series.py
@@ -2947,7 +2947,8 @@ class Series(np.ndarray):
         Returns:
             If vertices: tuple of grid mask and vertices_dict. Otherwise, grid mask only.
                 - grid mask: Series object with voxel=1 inside ROI.
-                  Series object with shape (nz,ny,nx) from original image,
+                  Series object with shape (nz,ny,nx) (3D or more) or
+                  (ny, nx) (2D) from original image,
                   dtype ubyte. Voxel inside ROI is 1, 0 outside.
                 - vertices_dict: if vertices: Dictionary of vertices.
 
@@ -3029,7 +3030,9 @@ class Series(np.ndarray):
             If vertices: tuple of grid mask and vertices_dict. Otherwise, grid mask only.
                 - grid mask: Series object with voxel=1 inside ROI.
                   Series object with shape (nz,ny,nx) from original image,
-                  dtype ubyte. Voxel inside ROI is 1, 0 outside.
+                  dtype ubyte. If original image is 2D, the mask
+                  will be shape (ny,nx) from original image.
+                  Voxel inside ROI is 1, 0 outside.
                 - vertices_dict: if vertices: Dictionary of vertices.
 
         Raises:
@@ -3051,7 +3054,10 @@ class Series(np.ndarray):
             if follow:
                 new_grid = np.zeros_like(self, dtype=np.ubyte)
             else:
-                new_grid = np.zeros((self.slices, self.rows, self.columns), dtype=np.ubyte)
+                if self.ndim == 2:
+                    new_grid = np.zeros((self.rows, self.columns), dtype=np.ubyte)
+                else:
+                    new_grid = np.zeros((self.slices, self.rows, self.columns), dtype=np.ubyte)
             new_grid = Series(new_grid, input_order=input_order, template=self, geometry=self)
         new_grid.seriesDescription = 'ROI'
         new_grid.windowCenter = .5