+++ CHANGELOG.md
@@ -7,11 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!--next-version-placeholder-->
 
-## [v3.4.2-dev0] - 2024-01-16
+## [v3.4.2-dev0] - 2024-01-18
 ### Changed
+* Series.get_roi_mask(): Return 2D mask when original images is 2D.
 * Viewer: Display image text even when some attributes are missing.
 * Viewer: Better presentation of window center/width when they are floats.
-* Viewer: Reimplemented PageDown/PageUp to scroll one row at a time.
+* Viewer: Reimplemented PageDown/PageUp to scroll one page at a time.
+* Viewer: Adjusting window center/width is modified to use
+  series min/max, not existing window.
+  Will improve user feedback when window is unreasonable.
+
+### Added
+* Viewer: Ctrl+Home/End will scroll to first/last series.
+* Viewer: Ctrl+Array Left/Right will scroll one series.
+* Viewer: Ctrl+Array Up/Down will scroll one row of series.
+* Viewer: Toggle the hide text will affect all series.
 
 ## [v3.4.1] - 2024-01-15
 ### Changed